feat: add AppConfigurator
